ShallowArray_GPU.h   1. Fix bug in ShallowArray_Flat: assignment operator= (for array)   2. Rename 'new_memory' -> new_memory_array; and add new 'new_memory_scalar' Class.C:   1. Fix bug: when using pointer in Shared, only allocate when the pointer is to an array
configure.py:
  1. generate CUDA-binary targetting multpiple GPUs (CC from 3.5 -> 7.5 )

VanDerPolCoupledSystem: add new model

+++ mdl/src/Class.C
@@ -1336,8 +1336,8 @@ void Class::addConstructor()
      fb << STR_GPU_CHECK_START;
      for (it = _attributes_gpu.begin(); it != end; ++it) {
        std::string attName = (*it)->getName();
-       if ((*it)->isPointer()) {
-	 fb << tab << attName  << " = new_memory<" << (*it)->getType() << ">();\n";
+       if ((*it)->isPointer() and (*it)->getType().find("ShallowArray") != std::string::npos) {
+	 fb << tab << attName  << " = new_memory_array<" << (*it)->getType() << ">();\n";
        }
      }
      fb << STR_GPU_CHECK_END;
@@ -1455,7 +1455,7 @@ void Class::addDestructor()
        fb << STR_GPU_CHECK_START;
        for (it = _attributes_gpu.begin(); it != end; ++it) {
 	 std::string attName = (*it)->getName();
-	 if ((*it)->isPointer()) {
+	 if ((*it)->isPointer() and (*it)->getType().find("ShallowArray") != std::string::npos) {
 	   fb << tab << "delete_memory(" << attName << ");\n";
 	 }
        }
